equipamentos/vending-machine/requirements.txt
python-dotenv
requests
python-socketio[client]
pyjwt
RPi.GPIO
